+++ backend-main/src/main/java/com/codecollab/oj/model/entity/QuestionUsecase.java
@@ -67,5 +67,5 @@ public class QuestionUsecase {
      * MB
      */
     @TableField(value = "memory_limit")
-    private Integer memoryLimit;
+    private Double memoryLimit;
 }
